Move CDBBase into seperated file for persistence
This file might be used for additional consolidation of database and persistence related blocks, as noted in the TODO comment block.
