+++ src/Makefile.am
@@ -161,6 +161,7 @@ OMNICORE_H = \
   mastercore_log.h \
   mastercore_mdex.h \
   mastercore_parse_string.h \
+  mastercore_persistence.h \
   mastercore_rpc.h \
   mastercore_script.h \
   mastercore_sp.h \
@@ -174,6 +175,7 @@ OMNICORE_CPP = \
   mastercore_log.cpp \
   mastercore_mdex.cpp \
   mastercore_parse_string.cpp \
+  mastercore_persistence.cpp \
   mastercore_rpc.cpp \
   mastercore_script.cpp \
   mastercore_sp.cpp \
