+++ src/mastercore.cpp
@@ -20,6 +20,7 @@
 #include "mastercore_errors.h"
 #include "mastercore_log.h"
 #include "mastercore_mdex.h"
+#include "mastercore_persistence.h"
 #include "mastercore_script.h"
 #include "mastercore_sp.h"
 #include "mastercore_tx.h"
